refactor
- use physical core num as default njob
  this will reduce the cost on lock in both userspace and kernel space
  a little slower, but more efficient
- refactor walkdir
- some other minor refactors

+++ src/executor.rs
@@ -18,7 +18,7 @@ pub struct Executor {
 }
 
 impl Executor {
-    fn new(nthreads: u32) -> Self {
+    fn new(nthreads: u8) -> Self {
         let (sender, receiver) = unbounded::<Runnable>();
         for i in 0..nthreads {
             let receiver = receiver.clone();
